Support multiple backends.
This commit separates Windows specific bits into a distinct module and
adds support for basic system call analysis for Linux-based VMs. The
appropriate backend is selected based on libvmi config. The Linux
backend currently associates system call numbers with their names using
the system call table and kernel symbol map.

Additionally, it is worth noting that these changes expect the latest
libvmi version from Git.

It is worth noting that the Windows backend is likely to be broken at
present as I still don't have a Windows VM for testing. However, the
fixes are likely to be trivial.

Everything is still very much in flux.