Tests that work to some extent.
Now that I got my testing rig working again, I continued implementing
the missing testing functionality. This version seems to mostly be
working, that is, the system is able to run the VMs, generate, mount,
and execute CDROMs and execute tests. However, there is still
something wrong with them (or the Linux analyzer) as the analyzer
crashes with weird memory access errors. My current hypothesis is that
it is somehow related to the way the system is turned off. This needs
further investigation.

There is still some fixing to do with the Windows tests.